% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/determine_n_t_anova.R
\name{determine_n_t_anova}
\alias{determine_n_t_anova}
\title{Determine N required to have a significant *t*-test or one-way ANOVA}
\usage{
determine_n_t_anova(
  groups,
  means,
  sds,
  alpha = 0.05,
  desired_power = 0.8,
  alternative = NULL,
  eq_var = NULL,
  mu = NULL,
  paired = F,
  anova_type = NULL,
  effect_size = F
)
}
\arguments{
\item{groups}{Positive Integer Scalar. The number of groups to perform the test with.}

\item{means}{Numeric Vector. The theoretical means for each group.}

\item{sds}{Positive Numeric Integer or Vector. The theoretical standard deviation(s) for the group(s).}

\item{alpha}{Positive Scalar between 0 and 1. The significance threshold for the test. Defaults to .05}

\item{desired_power}{Numeric (Float) between 0 and 1. The desired power of the test. Defaults to .8}

\item{alternative}{Character (String). The alternative hypothesis to be used by a t-test. Values of "two.sided" (default), "less", or "greater" are valid. Ignored if there are 3 or more groups.}

\item{eq_var}{Logical (Boolean). Should a two-sample t-test assume equal variance (i.e., perform Student's *t*-test). Defaults to FALSE. Ignored if only 1 group or more than 2.}

\item{mu}{Integer Scalar. Null mean to test against for a one-sample test. Defaults to 0. Ignored if groups are greater than 1.}

\item{paired}{Logical (Boolean). Should a two-sample t-test be conducted in a pairwise manner. Defaults to FALSE. Ignored if only 1 group or more than 2.}

\item{anova_type}{Integer Scalar from 1 to 3. The type of Sums of Squared Deviations to be used for the one-way ANOVA. Ignored if less than 3 groups.}

\item{effect_size}{Logical (Boolean). Should the standardized effect size (Cohen's *d* or *f2*) be returned? Defaults to `FALSE`}
}
\description{
See Title
}
