% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eta_sq.R
\name{eta_sq}
\alias{eta_sq}
\title{Calculate Eta-Squared for ANOVA}
\usage{
eta_sq(aov_in, type = "1")
}
\arguments{
\item{aov_in}{An object generated by the `aov()` function. This should be your omnibus ANOVA.}

\item{type}{Character. This indicates the type of Sums of Squares from which the omega-squared statistic should be calculated. Options are digits from 1 to 3 and the corresponding roman numerals. Defaults to type 1.}
}
\value{
Returns a `data.frame` containing all of the relevant information.
}
\description{
@description
 Currently only works for non-repeated measures ANOVA...
}
