% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fishers_z.R
\name{fishers_z}
\alias{fishers_z}
\title{Calculate Fisher's *z* from a correlation value}
\usage{
fishers_z(r, n = NULL, alpha = 0.05, return_se = F)
}
\arguments{
\item{r}{Numeric. The correlation to be converted to a *z*-score}

\item{n}{Numeric (Optional). if a sample size is provided, a confidence interval (defined by `alpha`) will be returned along with the *z* value}

\item{alpha}{The corresponding alpha level for the confidence interval to be returned if `n` is not `NULL`}

\item{return_se}{Logical. If `True`, the standard error of the *z*-score will be returned along side of the confidence interval. Only used if `n` is provided. Defaults to `False`.}
}
\value{
If `n` is not provided, the function will only return the calculated *z*-score. If `n` is provided, a list will be returned: the first item will be the *z*-score, the second item will be the lower-bound of the confidence interval computed, the third will be the corresponding upper-bound. An optional 4th value is the standard error.
}
\description{
`fishers_z` is a function designed to take correlation values and return a Fisher's *z* form that value.
}
