% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hierarchical_reg.R
\name{hierarchical_reg}
\alias{hierarchical_reg}
\title{A Function for Step-wise & Hierarchical Regression}
\usage{
hierarchical_reg(
  y,
  chunks = NULL,
  data = NULL,
  stepwise = F,
  verbose = F,
  steps_verbose = NULL,
  step_iters = 50
)
}
\arguments{
\item{y}{Either: Character or an object created by the `lm()` function. If this is argument is a character string it should specify the outcome variable for the model. If it is a linear model object, it should be the full model with the outcome regressed on all predictors including those at each step.}

\item{chunks}{List of Character Vectors. This list should specify the predictors in the model that should be chunked together.
At the highest level the list should be specified using the `list()` method, but each vector should be specified using the `c()` method.
For example, one version of the argument could be `chunks = list(step1=c("pred1", "pred2"), step2=c("pred3"))`. If step names are not specified, the chunks will be numbered in the order that they are entered into the model.
If `chunks` are not specified, the individual predictors will be treated as the items to be included at each step.}

\item{data}{Data.frame. The data from which the model(s) should be calculated. This argument is necessary if `y` is a character string.}

\item{stepwise}{Logical. If False (default), the `chunks` will be entered into the model in the order that they are listed. If True, the `chunks` will be entered into the model based on the amount of variance they explain.}

\item{verbose}{Logical. If `True` the coefficients of the model will be output to the console for each chunk bring processed while the function runs. This will not be saved, as it is already part of the `lm` object and cam be accessed from there. Defaults to `False`.}

\item{steps_verbose}{Logical. If `True` the order that the chunks were entered into the model will be output to the console while the function tries to identify the optimal order. If `NULL` (default) this will will take the value of `verbose`.}

\item{step_iters}{Numeric. Only used if `stepwise == True`. The number of iterations of step-wise comparisons to do before ending the procedure comparing variances explained. Defaults to 50.}
}
\value{
A list of data.frames. The first is the information about the different steps/chunks and their R-squared Changes. The second is the list of coefficient summaries at each step.
}
\description{
A Function for Step-wise & Hierarchical Regression
}
