% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mediation.R
\name{lm_mediation}
\alias{lm_mediation}
\title{Simple Mediation analyses}
\usage{
lm_mediation(mediator, y_xm_reg, std = T)
}
\arguments{
\item{mediator}{Character. The name of the mediator to be modeled.}

\item{y_xm_reg}{An object generated from the `lm()` function. This regression should be for the outcome variable regressed on both the primary predictor and the mediator.}

\item{std}{Logical. If `True` (default) the function will output rows for both the unstandardized and standardized mediation effects.}
}
\value{
Returns a `data.frame` with the mediation effect and test of significance.
}
\description{
The `psych::mediate()` function is recommended over this, as, at present, this function
handles only simple mediation models (i.e., one IV and one mediator). The function from the
`psych` package is able to much more sophisticated models and will output a diagram for you.
}
