% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/omega_squared.R
\name{omega_sq}
\alias{omega_sq}
\title{Calculate Omega Squared for ANOVA tables}
\usage{
omega_sq(aov_in, type = c("I", "II", "III", "1", "2", "3"))
}
\arguments{
\item{aov_in}{An object generated by the `aov()` function. This sohuld be your omnibus ANOVA.}

\item{type}{Character. This indicated the type of Sums of Squares from which the omega-squared statistic should be calculated. Options are digits from 1 to 3 and the corresponding roman numerals.}
}
\value{
Returns a `data.frame` containing all of the relevant information.
}
\description{
Calculate Omega Squared for ANOVA tables
}
