% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pairwise_post_hocs_anova.R
\name{pairwise_post_hocs_anova}
\alias{pairwise_post_hocs_anova}
\title{Pairwise Comparisons for ANOVA}
\usage{
pairwise_post_hocs_anova(
  aov_in,
  ss_type = "1",
  comp_type = c("Tukey", "LSD", "Bonferroni")
)
}
\description{
Not functional yet. But hopefully it soon will be...
}
