% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_pow_anova.R
\name{post_hoc_pow_anova}
\alias{post_hoc_pow_anova}
\title{Calculates Post-hoc Power for ANOVA}
\usage{
post_hoc_pow_anova(aov_in, type = "1", alpha = 0.05, f2 = F, omega = F)
}
\arguments{
\item{aov_in}{An object generated by the `aov()` function. This should be your omnibus ANOVA.}

\item{type}{Character (String). This indicates the type of Sums of Squares from which the power should be calculated. Options are digits from 1 to 3 and the corresponding roman numerals. Defaults to type 1.}

\item{f2}{Logical (Boolean). Sould the function also return the Cohen's f^2 effect sizes? Defaults to false.}

\item{omega}{Logical (Boolean). Should omega-squared be used to calculate power? Defaults to `FALSE`, indicating (partial) eta-squared will be used.}
}
\description{
see title
}
