% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_sim_pow_anova.R
\name{post_hoc_sim_pow_anova}
\alias{post_hoc_sim_pow_anova}
\title{Simulate Power for ANOVA}
\usage{
post_hoc_sim_pow_anova(
  aov_in,
  type,
  n_forSim = NULL,
  n_sims = 10000,
  alpha = 0.05
)
}
\arguments{
\item{aov_in}{An `aov()` object. The ANOVA model for which to simulate power.}

\item{n_forSim}{Integer. The number of observations to be used in the linear model simulations. Defaults to the n (and ratios between groups) of the original data from the `aov_in`}

\item{n_sims}{Integer. The number of simulations to run. Defaults to 10k.}

\item{alpha}{Float between 0 and 1. The significance level of the test.}
}
\description{
see title
}
