% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_sim_pow_lm.R
\name{post_hoc_sim_pow_lm}
\alias{post_hoc_sim_pow_lm}
\title{Simulate Power for A Linear Model}
\usage{
post_hoc_sim_pow_lm(lm, n_forSim = NULL, n_sims = 10000, alpha = 0.05)
}
\arguments{
\item{lm}{An `lm()` object. The regression model for which to simulate power.}

\item{n_forSim}{Integer. The number of observations to be used in the linear model simulations.}

\item{n_sims}{Integer. The number of simulations to run.}

\item{alpha}{Float between 0 and 1. The significance level of the test.}
}
\description{
see title
}
