% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pow_t_anova.R
\name{pow_t_anova}
\alias{pow_t_anova}
\title{Calculate power for *t*-test or One-Way Anova}
\usage{
pow_t_anova(
  groups,
  n,
  means,
  sds,
  alpha = 0.05,
  alternative = NULL,
  eq_var = NULL,
  mu = NULL,
  paired = F,
  anova_type = NULL,
  show_plot = F
)
}
\arguments{
\item{groups}{Positive Integer Scalar. The number of groups to perform the test with.}

\item{n}{Positive Integer Scalar or Vector. The number of observations per group. If a scalar is provided, all groups will have that number of observations. If a vector is provided, the there must be a number of integers equal to `groups`; these will be the numbers of observations in a given group.}

\item{means}{Numeric Vector. The theoretical means for each group.}

\item{sds}{Positive Numeric Integer or Vector. The theoretical standard deviation(s) for the group(s).}

\item{alpha}{Positive Scalar between 0 and 1. The significance threshold for the test. Defaults to .05}

\item{alternative}{Character (String). The alternative hypothesis to be used by a t-test. Values of "two.sided" (default), "less", or "greater" are valid. Ignored if there are 3 or more groups.}

\item{eq_var}{Logical (Boolean). Should a two-sample t-test assume equal variance (i.e., perform Student's *t*-test). Defaults to FALSE. Ignored if only 1 group or more than 2.}

\item{mu}{Integer Scalar. Null mean to test against for a one-sample test. Defaults to 0. Ignored if groups are greater than 1.}

\item{paired}{Logical (Boolean). Should a two-sample t-test be conducted in a pairwise manner. Defaults to FALSE. Ignored if only 1 group or more than 2.}

\item{anova_type}{Integer Scalar from 1 to 3. The type of Sums of Squared Deviations to be used for the one-way ANOVA. Ignored if less than 3 groups.}

\item{n_sims}{Positive Integer Scalar. The number of simulations to base the estimate on. Default is 10,000.}
}
\description{
Calculates power for a given means, sd, sample size(s), and alpha-level.
}
