% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat_anova.R
\name{sim_dat_anova}
\alias{sim_dat_anova}
\title{Simulate new data based on an extant ANOVA}
\usage{
sim_dat_anova(aov_in, n_forSim = NULL)
}
\arguments{
\item{aov_in}{Object of type `aov`. The object from which to generate a new data-set.}

\item{n_forSim}{Integer Scalar. The number of observations to create in the new data. If `NULL` the same number of observations as in the original `lm` will be used.}
}
\description{
see title
}
