% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat_lm.R, R/sim_pow_lm.R
\name{sim_dat_lm}
\alias{sim_dat_lm}
\title{Simulate new data based on a Linear Regression Model}
\usage{
sim_dat_lm(lm, n_forSim = NULL)

sim_dat_lm(lm, n_forSim = NULL)
}
\description{
see title

see title
}
