% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pow_t_z_anova.R
\name{sim_dat_t_anova}
\alias{sim_dat_t_anova}
\title{Simulated power tests for a single Categorical IV and a Continuous DV}
\usage{
sim_dat_t_anova(groups, n, means, sds)
}
\arguments{
\item{groups}{Positive Integer Scalar. The number of groups to perform the test with.}

\item{n}{Positive Integer Scalar or Vector. The number of observations per group. If a scalar is provided, all groups will have that number of observations. If a vector is provided, the there must be a number of integers equal to `groups`; these will be the numbers of observations in a given group.}

\item{means}{Numeric Vector. The theoretical means for each group.}

\item{sds}{Positive Numeric Integer or Vector. The theoretical standard deviation(s) for the group(s).}

\item{n_sims}{Positive Integer Scalar. The number of simulations to base the estimate on. Default is 10,000.}

\item{alpha}{Positive Scalar between 0 and 1. The significance threshold for the test. Defaults to .05}

\item{alternative}{Character (String). The alternative hypothesis to be used by a t-test. Values of "two.sided" (default), "less", or "greater" are valid. Ignored if there are 3 or more groups.}

\item{mu}{Integer Scalar. Null mean to test against for a one-sample test. Defaults to 0. Ignored if groups are greater than 1.}

\item{eq_var}{Logical (Boolean). Should a two-sample t-test assume equal variance (i.e., perform Student's *t*-test). Defaults to FALSE. Ignored if only 1 group or more than 2.}

\item{paired}{Logical (Boolean). Should a two-sample t-test be conducted in a pairwise manner. Defaults to FALSE. Ignored if only 1 group or more than 2.}

\item{anova_type}{Integer Scalar from 1 to 3. The type of Sums of Squared Deviations to be used for the one-way ANOVA. Ignored if less than 3 groups.}
}
\description{
These functions will estimate power via simulation for one- and two-sample *t*-tests, *z*-test, and one-way ANOVA.
As these tests assume a normal distribution from the population, all simulated samples are drawn from such.
}
