% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pow_lmer.R
\name{sim_pow_lmer}
\alias{sim_pow_lmer}
\title{Power Analysis via Simulation for Linear Mixed Effects Model}
\usage{
sim_pow_lmer(J = 68, K = 32, n.sims = 10000, cor.PnR.supp = 0.21)
}
\description{
NOT YET FUNCTIONAL
}
