% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_dependent_cors.R
\name{z_dependent_cors}
\alias{z_dependent_cors}
\title{Calculate the *z*-value comparing two dependent correlations}
\usage{
z_dependent_cors(r_x1_y, r_x2_y, r_x1_x2, n)
}
\arguments{
\item{r_x1_y}{Numeric. The correlation value between the first predictor and the outcome.}

\item{r_x2_y}{Numeric. The correlation value between the second predictor and the outcome.}

\item{r_x1_x2}{Numeric. The correlation value between the first and second predictors.}

\item{n}{Numeric. The sample size used to calculate the correlations.}
}
\value{
Returns a z-score that can be tested using the `z_dependent_cors_test()` or `z_test()` function.
}
\description{
Calculate the *z*-value comparing two dependent correlations
}
