% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_dependent_cors_test.R
\name{z_dependent_cors_test}
\alias{z_dependent_cors_test}
\title{Steiger's z*-bar (Steiger, 1980)}
\usage{
z_dependent_cors_test(
  r_x1_y = NULL,
  r_x2_y = NULL,
  r_x1_x2 = NULL,
  n = NULL,
  z_val = NULL,
  alpha = 0.05,
  two_tailed = T
)
}
\arguments{
\item{r_x1_y}{Numeric. The correlation value between the first predictor and the outcome.}

\item{r_x2_y}{Numeric. The correlation value between the second predictor and the outcome.}

\item{r_x1_x2}{Numeric. The correlation value between the first and second predictors.}

\item{n}{Numeric. The sample size used to calculate the correlations.}

\item{z_val}{Numeric. A z-score to be tested for significance.}

\item{alpha}{Numeric. A confidence level for the test. Defaults to .05, corresponding with a 95\% confidence level.}

\item{two_tailed}{Logical. If `True` (default) the test for significance is done in a two-tailed manner.}
}
\value{
Outputs the test for significance to the console and returns both a *z*-value and *p*-value for the test.
}
\description{
This code allows you to test for a significant difference in the correlations between two variables and a third.
}
