% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_independent_cors.R
\name{z_independent_cors}
\alias{z_independent_cors}
\title{Calculate the statistic for a test of significant differences between two independent correlations.}
\usage{
z_independent_cors(r1, r2, n1, n2)
}
\arguments{
\item{r1}{Numeric. The first correlation to be compared. Must be between -1 and 1.}

\item{r2}{Numeric. The second correlation to be compared. Must be between -1 and 1.}

\item{n1}{Numeric. The sample size from which `r1` was generated. Must be a positive value.}

\item{n2}{Numeric. The sample size from which `r2` was generated. Must be a positive value.}
}
\value{
`z_independent_cors()` returns a z-score that can be tested using the `z_independent_cors_test()` or `z_test()` function.
}
\description{
Calculate the statistic for a test of significant differences between two independent correlations.
}
