% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_test.R
\name{z_test}
\alias{z_test}
\title{A simple method for outputting the results of a *z*-test}
\usage{
z_test(z_val = NULL, alpha = 0.05, two_tailed = T)
}
\arguments{
\item{z_val}{Numeric. The *z*-value to be tested for statistical significance.}

\item{alpha}{Numeric. The significance level criterion for the test. Must be between 0 and 1. Defaults to .05, corresponding with a confidence level of 95\%.}

\item{two_tailed}{Logical. If `True` (defualt) the test will be conducted in a two-tailed manner. If `False`, the test is conducted in a one-tailed manner. (Indication of a specific directional hypothesis is not currently supported.)}
}
\value{
returns a list containing the `z_val` initially provided and the corresponding `p_val` (i.e., the probability of obtaining the specified value, or one more extreme, assuming the null hypothesis is true). The function also prints out the results of the test.
}
\description{
A simple method for outputting the results of a *z*-test
}
