% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/post_hoc_sim_pow_lm.R
\name{post_hoc_sim_pow_lm}
\alias{post_hoc_sim_pow_lm}
\title{Simulate Power for A Linear Model}
\usage{
post_hoc_sim_pow_lm(lm, n_forSim = NULL, n_sims = 10000, alpha = 0.05)
}
\description{
see title
}
