% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat_lm.R, R/sim_pow_lm.R
\name{sim_dat_lm}
\alias{sim_dat_lm}
\title{Simulate new data based on a Linear Regression Model}
\usage{
sim_dat_lm(lm, n_forSim = NULL)

sim_dat_lm(lm, n_forSim = NULL)
}
\arguments{
\item{lm}{Object of type `lm`. The object from which to generate a new data-set.}

\item{n_forSim}{Integer Scalar. The number of observations to create in the new data. If `NULL` the same number of observations as in the original `lm` will be used.}
}
\description{
see title

see title
}
