% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_dat_t_anova.R
\name{sim_dat_t_anova}
\alias{sim_dat_t_anova}
\title{Simulated power tests for a single Categorical IV and a Continuous DV}
\usage{
sim_dat_t_anova(groups, n, means, sds)
}
\arguments{
\item{groups}{Positive Integer Scalar. The number of groups to perform the test with.}

\item{n}{Positive Integer Scalar or Vector. The number of observations per group. If a scalar is provided, all groups will have that number of observations. If a vector is provided, the there must be a number of integers equal to `groups`; these will be the numbers of observations in a given group.}

\item{means}{Numeric Vector. The theoretical means for each group.}

\item{sds}{Positive Numeric Integer or Vector. The theoretical standard deviation(s) for the group(s).}
}
\description{
This function will generate data for one- and two-sample *t*-tests, *z*-test, and one-way ANOVA.
As these tests assume a normal distribution from the population, all simulated samples are drawn from such.
}
